+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/rule/bestpractices/UnitTestShouldIncludeAssertRule.java
@@ -5,13 +5,18 @@
 package net.sourceforge.pmd.lang.java.rule.bestpractices;
 
 import java.util.Set;
+import java.util.function.Predicate;
 import java.util.stream.Collectors;
 
+import net.sourceforge.pmd.lang.java.ast.ASTAnnotation;
 import net.sourceforge.pmd.lang.java.ast.ASTBlock;
+import net.sourceforge.pmd.lang.java.ast.ASTClassLiteral;
 import net.sourceforge.pmd.lang.java.ast.ASTMethodCall;
 import net.sourceforge.pmd.lang.java.ast.ASTMethodDeclaration;
+import net.sourceforge.pmd.lang.java.ast.ASTTypeDeclaration;
 import net.sourceforge.pmd.lang.java.rule.AbstractJavaRulechainRule;
 import net.sourceforge.pmd.lang.java.rule.internal.TestFrameworksUtil;
+import net.sourceforge.pmd.lang.java.types.TypeTestUtil;
 import net.sourceforge.pmd.properties.PropertyDescriptor;
 import net.sourceforge.pmd.properties.PropertyFactory;
 
@@ -31,16 +36,32 @@ public class UnitTestShouldIncludeAssertRule extends AbstractJavaRulechainRule {
 
     @Override
     public Object visit(ASTMethodDeclaration method, Object data) {
-        ASTBlock body = method.getBody();
+        boolean usesSoftAssertExtension = usesSoftAssertExtension(method.getEnclosingType());
         Set<String> extraAsserts = getProperty(EXTRA_ASSERT_METHOD_NAMES);
+        Predicate<ASTMethodCall> isAssertCall = usesSoftAssertExtension
+                ? TestFrameworksUtil::isSoftAssert
+                : TestFrameworksUtil::isProbableAssertCall;
+
+        ASTBlock body = method.getBody();
         if (body != null
             && TestFrameworksUtil.isTestMethod(method)
             && !TestFrameworksUtil.isExpectAnnotated(method)
             && body.descendants(ASTMethodCall.class)
-                   .none(call -> TestFrameworksUtil.isProbableAssertCall(call)
-                           || extraAsserts.contains(call.getMethodName()))) {
+                .none(isAssertCall
+                        .or(call -> extraAsserts.contains(call.getMethodName())))) {
             asCtx(data).addViolation(method);
         }
         return data;
+    }
+
+    private boolean usesSoftAssertExtension(ASTTypeDeclaration typeDeclaration) {
+        ASTAnnotation extendWith = typeDeclaration.getAnnotation("org.junit.jupiter.api.extension.ExtendWith");
+        if (extendWith == null) {
+            return false;
+        }
+        return extendWith.getFlatValue("value")
+                    .filterIs(ASTClassLiteral.class)
+                    .map(ASTClassLiteral::getTypeNode)
+                    .any(c -> TypeTestUtil.isA("org.assertj.core.api.junit.jupiter.SoftAssertionsExtension", c));
     }
 }