+++ docs/pages/release_notes.md
@@ -22,6 +22,8 @@ This is a {{ site.pmd.release_type }} release.
 * java
   * [#5293](https://github.com/pmd/pmd/issues/5293): \[java] Deadlock when executing PMD in multiple threads
   * [#5324](https://github.com/pmd/pmd/issues/5324): \[java] Issue with type inference of nested lambdas
+* java-bestpractices
+  * [#4113](https://github.com/pmd/pmd/issues/4113): \[java] JUnitTestsShouldIncludeAssert - false positive with SoftAssertionsExtension
 
 ### 🚨 API Changes